+++ example/lib/pages/basic_scroll_page.dart
@@ -1,4 +1,3 @@
-import 'dart:collection';
 import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:silky_scroll/silky_scroll.dart';