+++ lib/src/silky_scroll_web_helper/silky_scroll_non_web_helper.dart
@@ -8,4 +8,9 @@ class SilkyScrollWebManager implements SilkyScrollWebManagerInterface {
   void blockOverscrollBehaviorXHtml() {
     // No-op on non-web platforms.
   }
+
+  @override
+  void setOverscrollBehaviorX(OverscrollBehaviorX value) {
+    // No-op on non-web platforms.
+  }
 }